+++ src/commands/account/outgoing.ts
@@ -35,9 +35,9 @@ export default class extends AccountTransactionsCommand {
         this.spinner.start();
 
         const profile = this.getProfile(options);
-        const accountHttp =  new AccountHttp(profile.url);
         const address = new AddressResolver().resolve(options, profile);
 
+        const accountHttp =  new AccountHttp(profile.url);
         accountHttp.getAccountOutgoingTransactions(address, options.getQueryParams())
             .subscribe((transactions) => {
                 this.spinner.stop(true);
@@ -48,10 +48,8 @@ export default class extends AccountTransactionsCommand {
                 console.log(text === '' ? 'There aren\'t outgoing transaction' : text);
             }, (err) => {
                 this.spinner.stop(true);
-                let text = '';
-                text += chalk.red('Error');
                 err = err.message ? JSON.parse(err.message) : err;
-                console.log(text, err.body && err.body.message ? err.body.message : err);
+                console.log(chalk.red('Error'), err.body && err.body.message ? err.body.message : err);
             });
     }
 }